+++ client/src/components/LayersControl.tsx
@@ -1,5 +1,6 @@
 import {
   Button,
+  Checkbox,
   FormControl,
   IconButton,
   Input,
@@ -25,12 +26,25 @@ import { useDropzone } from "react-dropzone";
 import AutoSizer from "react-virtualized-auto-sizer";
 import { formatElapsedTime } from "../utils/helpers";
 import { useConfirm } from "material-ui-confirm";
+import { area as turfArea } from "@turf/turf";
+
+import ErrorIcon from "@mui/icons-material/Error";
+import EditIcon from "@mui/icons-material/Edit";
 
 const LayersControl: FC = () => {
+  const minimumValidArea = useStore((state) => state.minimumValidArea);
   const [activeJob, setActiveJob] = useStore((state) => [state.activeJob, state.setActiveJob]);
 
   const [multipolygons, setMultipolygons] = useStore((state) => [state.multipolygons, state.setMultipolygons]);
   const [loadedGeoJSON, setLoadedGeoJSON] = useStore((state) => [state.loadedGeoJSON, state.setLoadedGeoJSON]);
+
+  const loadedGeoJSONArea = useMemo(() => {
+    if (!loadedGeoJSON) {
+      return 0;
+    }
+
+    return turfArea(loadedGeoJSON);
+  }, [loadedGeoJSON]);
 
   const [rows, setRows] = useStore((state) => [state.locations, state.setLocations]);
   const visibleLayerCount = useMemo(() => {
@@ -49,6 +63,9 @@ const LayersControl: FC = () => {
   const closeNewJob = useStore((state) => state.closeNewJob);
   const prepareGeoJSON = useStore((state) => state.prepareGeoJSON);
 
+  const toggleARDTiles = useStore((state) => state.toggleARDTiles);
+  const showARDTiles = useStore((state) => state.showARDTiles);
+
   const userInfo = useStore((state) => state.userInfo);
   const canWriteJobs = useMemo(() => userInfo?.permissions.includes("write:jobs"), [userInfo?.permissions]);
 
@@ -60,6 +77,27 @@ const LayersControl: FC = () => {
   const canSubmitBulkJob = useMemo(() => {
     return jobName && loadedFile && multipolygons.length > 0 && startYear <= endYear;
   }, [jobName, loadedFile, multipolygons, startYear, endYear]);
+
+  const isBulkJob = useMemo(() => {
+    return canSubmitBulkJob && multipolygons.length > 1;
+  }, [canSubmitBulkJob, multipolygons]);
+
+  const isValidArea = useMemo(() => {
+    // Landsat resolution is 30m, so we want to make sure the area is at least 900m^2
+    if (isBulkJob) {
+      let allRowsValid = rows.every((row) => {
+        if (!row.visible) {
+          return true;
+        }
+
+        return row.isValidArea;
+      });
+
+      return allRowsValid && visibleLayerCount > 0;
+    } else {
+      return loadedGeoJSONArea >= minimumValidArea;
+    }
+  }, [isBulkJob, loadedGeoJSONArea, multipolygons, rows, minimumValidArea, visibleLayerCount]);
 
   const validYears = useMemo(() => {
     return Array.from({ length: maxYear - minYear + 1 }, (_, i) => minYear + i);
@@ -79,21 +117,29 @@ const LayersControl: FC = () => {
         long = geojson?.features?.[0]?.geometry?.coordinates[0][0][1];
       }
 
+      let area = geojson?.properties?.shape_Area;
+      if (!area) {
+        area = turfArea(geojson);
+      }
+
+      let isValidArea = area > minimumValidArea;
+
       return {
-        visible: true,
+        visible: isValidArea,
         name: name,
         acres: geojson?.properties?.Acres,
         comments: geojson?.properties?.Comments,
         county: geojson?.properties?.County,
         polygon_So: geojson?.properties?.Polygon_So,
-        shape_Area: geojson?.properties?.Shape_Area,
-        shape_Leng: geojson?.properties?.Shape_Leng,
+        shapeArea: area,
+        shapeLeng: geojson?.properties?.Shape_Leng,
         source: geojson?.properties?.Source,
         wUR_Basin: geojson?.properties?.WUR_Basin,
         id: index,
         lat: lat,
         long: long,
         crop: geojson?.properties?.CDL_Crop || "",
+        isValidArea: isValidArea,
       };
     });
 
@@ -161,6 +207,9 @@ const LayersControl: FC = () => {
 
       let roundedAcres = row?.acres ? Math.round(row.acres * 100) / 100 : "NaN";
 
+      let [editRowName, setEditRowName] = useState(false);
+      let [rowName, setRowName] = useState(row.name);
+
       return (
         <div key={row.id} className={`layer-row ${row.id === activeRowId ? "active" : ""}`} style={style}>
           <div className="left-btns">
@@ -190,7 +239,21 @@ const LayersControl: FC = () => {
           </div>
           <div
             className="details"
-            onClick={() => {
+            onClick={(evt) => {
+              // If has class ignore-select, don't do anything
+              if (evt.target instanceof HTMLElement && evt.target.classList.contains("ignore-select")) {
+                return;
+              }
+
+              // Also check if parent has ignore-select
+              if ((evt.target as any)?.parentElement?.classList.contains("ignore-select")) {
+                return;
+              }
+
+              if (evt.target instanceof HTMLInputElement) {
+                return;
+              }
+
               if (row.id === activeRowId) {
                 setActiveRowId(null);
                 setLoadedGeoJSON(null);
@@ -208,18 +271,54 @@ const LayersControl: FC = () => {
               }
             }}
           >
-            <Typography
-              variant="body1"
-              sx={{
-                color: row.id === activeRowId ? "var(--st-gray-10)" : "var(--st-gray-30)",
-                fontWeight: "bold",
-                display: "flex",
-                alignItems: "center",
-                cursor: "pointer",
-              }}
-            >
-              {row.name}
-            </Typography>
+            {!editRowName && (
+              <Typography
+                variant="body1"
+                className="ignore-select"
+                sx={{
+                  color: row.id === activeRowId ? "var(--st-gray-10)" : "var(--st-gray-30)",
+                  fontWeight: "bold",
+                  display: "flex",
+                  alignItems: "center",
+                  cursor: "pointer",
+                }}
+                onDoubleClick={() => {
+                  setEditRowName(true);
+                }}
+              >
+                {row.name}
+                <EditIcon
+                  onClick={(evt) => {
+                    evt.stopPropagation();
+                    setEditRowName(true);
+                  }}
+                  className="ignore-select"
+                  sx={{ color: "var(--st-gray-30)", cursor: "pointer", ml: "8px", fontSize: "14px" }}
+                />
+              </Typography>
+            )}
+            {editRowName && (
+              <>
+                <input
+                  placeholder={row.name}
+                  value={rowName}
+                  style={{
+                    backgroundColor: "var(--st-gray-90)",
+                    border: "1px solid var(--st-gray-40)",
+                    borderRadius: "4px",
+                    padding: "4px",
+                  }}
+                  onChange={(evt) => {
+                    setRowName(evt.target.value);
+                  }}
+                  onBlur={() => {
+                    row.name = rowName;
+                    setRows([...rows]);
+                    setEditRowName(false);
+                  }}
+                />
+              </>
+            )}
             <Typography
               variant="body2"
               style={{ color: row.id === activeRowId ? "var(--st-gray-20)" : "var(--st-gray-40)" }}
@@ -229,9 +328,15 @@ const LayersControl: FC = () => {
             {row?.acres && (
               <Typography
                 variant="body2"
-                style={{ color: row.id === activeRowId ? "var(--st-gray-20)" : "var(--st-gray-40)" }}
+                style={{
+                  color: row.isValidArea
+                    ? row.id === activeRowId
+                      ? "var(--st-gray-20)"
+                      : "var(--st-gray-40)"
+                    : "var(--st-red)",
+                }}
               >
-                Acres: {roundedAcres}
+                Acres: {roundedAcres} {row.isValidArea ? "" : "(Area too small)"}
               </Typography>
             )}
           </div>
@@ -345,7 +450,7 @@ const LayersControl: FC = () => {
             sx={{ marginLeft: "auto" }}
             onClick={() => {
               rows.forEach((row) => {
-                row.visible = true;
+                row.visible = row.isValidArea;
               });
               setRows([...rows]);
             }}
@@ -394,10 +499,7 @@ const LayersControl: FC = () => {
           </AutoSizer>
         </div>
       )}
-      <div
-        className="dropzone-container"
-        style={{ height: !loadedFile && rows.length === 0 ? 400 : 200, width: "300px" }}
-      >
+      <div className="dropzone-container" style={{ height: !loadedFile && rows.length === 0 ? 400 : 200, width: "318px" }}>
         {loadedFile && (
           <div className="cancel-job">
             <IconButton
@@ -412,11 +514,7 @@ const LayersControl: FC = () => {
             </IconButton>
           </div>
         )}
-        <div
-          className={`dropzone-area ${isDragActive ? "drag-active" : ""}`}
-          {...getRootProps()}
-          style={{ padding: "8px" }}
-        >
+        <div className={`dropzone-area ${isDragActive ? "drag-active" : ""}`} {...getRootProps()} style={{ padding: "8px" }}>
           <input {...getInputProps()} />
           {loadedFile ? (
             <div className="loaded-file" style={{ margin: "8px" }}>
@@ -448,13 +546,23 @@ const LayersControl: FC = () => {
         className="bottom-buttons"
         style={{ width: "100%", display: "flex", alignItems: "center", justifyContent: "flex-end" }}
       >
+        <div style={{ display: "flex", alignItems: "center", marginRight: "auto" }}>
+          <Checkbox
+            onClick={() => toggleARDTiles()}
+            checked={showARDTiles}
+            style={{ padding: 0, marginRight: "4px", marginLeft: "4px" }}
+          />
+          <Typography variant="body2" style={{ color: "var(--st-gray-30)", fontSize: "12px" }}>
+            Show Valid Bounds
+          </Typography>
+        </div>
         <Button
-          disabled={!canSubmitJob && !canSubmitBulkJob}
+          disabled={!isValidArea || (!canSubmitJob && !canSubmitBulkJob)}
           variant="contained"
-          color="primary"
+          color={isValidArea ? "primary" : "error"}
           style={{
             margin: "8px",
-            border: canSubmitJob || canSubmitBulkJob ? "1px solid #1E40AF" : "1px solid transparent",
+            border: isValidArea && (canSubmitJob || canSubmitBulkJob) ? "1px solid #1E40AF" : "1px solid transparent",
           }}
           onClick={() => {
             if (canSubmitJob) {
@@ -484,8 +592,19 @@ const LayersControl: FC = () => {
             }
           }}
         >
-          Submit {canSubmitBulkJob && multipolygons.length > 1 && "Bulk "}Job
+          Submit {isBulkJob && "Bulk "}Job
         </Button>
+      </div>
+      <div style={{ display: "flex", alignItems: "center", justifyContent: "center", marginBottom: "4px" }}>
+        {loadedGeoJSONArea > 0 && !isValidArea && (
+          <div style={{ display: "flex", alignItems: "flex-end", gap: "4px" }}>
+            <ErrorIcon style={{ color: "var(--st-red)", fontSize: "32px" }} />
+            <Typography variant="body2" className="note" style={{ fontSize: "12px", color: "var(--st-gray-40)" }}>
+              Area is too small (min {minimumValidArea} m<sup>2</sup>)<br />
+              Please upload a larger area.
+            </Typography>
+          </div>
+        )}
       </div>
     </div>
   );