+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+## 1.5.0 (2024-12-18)
+
+### Features
+- Full page report layout
+  - The report is now the size of a standard 8.5 x 11 inch page
+  - Avg. Cloud Coverage & Missing Data is now displayed as a separated line chart below the ET and PET chart
+  - ROI tiles are now bigger
+  - Updated wording for data sources at the bottom of the page
+- Precipitation data is now displayed in the report
+  - Precipitation data is displayed as a line chart below the ET and PET chart
+  - Data comes from the Oregon State PRISM dataset and is credited at the bottom of the generated report
+
+### Bug Fixes
+- Increased error handling for missing data in the report
+- Fixed an issue with certain variable tiles after 2018 not being correctly mapped to their relative S3 location
+
+## 1.4.0 (2024-12-11)
+
+### Features
+- Shows an error messsage and prevents the user from submitting a job if the target area is too small (less than 900 m^2)
+- Adds a toggle to show "valid bounds" on the map for the area we have data for
+- Shows all available properties in job status
+
+### Bug Fixes
+- Adds error handling for jobs involving areas that are too small being run
+  - The job would run as normal, fetch all necessary tiles, but then fail due to the "valid" area mask being empty as it didn't cover any pixels
+
 ## 1.3.0 (2024-12-05)
 
 ### Features